Correcting VideoTs type

+++ erizo/src/erizo/media/MediaProcessor.cpp
@@ -72,7 +72,7 @@ namespace erizo {
   int InputProcessor::deliverVideoData(char* buf, int len) {
     if (videoUnpackager && videoDecoder) {
       int estimatedFps=0;
-      int videoTs2 = 0;
+      double videoTs2 = 0;
       int ret = unpackageVideo(reinterpret_cast<unsigned char*>(buf), len,
           unpackagedBuffer_, &gotUnpackagedFrame_, &estimatedFps, &videoTs2);
       if (ret < 0)
